requirements.txt
streamlit
google-genai
streamlit-sortables